Add biome::rest_api::actix::register module
Adds a module intended to hold the actix implemenation of the
biome register route, currently behind the biome-credentials
and rest-api-actix feature flags.
