+++ libsplinter/src/biome/rest_api/mod.rs
@@ -64,10 +64,10 @@ pub use config::{BiomeRestConfig, BiomeRestConfigBuilder};
 pub use error::BiomeRestResourceManagerBuilderError;
 
 #[cfg(all(feature = "biome-credentials", feature = "rest-api-actix"))]
-use self::actix::login::make_login_route;
+use self::actix::{login::make_login_route, register::make_register_route};
 #[cfg(feature = "biome-credentials")]
 use super::credentials::{
-    rest_resources::{make_list_route, make_register_route, make_user_routes},
+    rest_resources::{make_list_route, make_user_routes},
     store::diesel::SplinterCredentialsStore,
 };
